+++ lunaengine/__init__.py
@@ -2,8 +2,8 @@
 LunaEngine - A powerful 2D game engine for Python
 """
 
-__version__ = "0.2.6"
-__status__ = "Alpha"
+__version__ = "0.2.6.2"
+__status__ = "Beta"
 
 from . import core, misc, ui, graphics, utils, backend, tools, storage